Qr genrater prokect.html/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.3",
    "pyjwt>=2.12.1",
    "qrcode[pil]>=8.2",
    "werkzeug>=3.1.8",
]

